fix: Correct random() usage in TourParticipant seed data
Convert Double range to Int range before calling random() to fix compilation error.
ClosedRange<Double> doesn't have random() extension function, only IntRange does.

+++ shared/src/commonMain/kotlin/com/lazytravel/data/models/TourParticipant.kt
@@ -77,7 +77,7 @@ data class TourParticipant(
         // Tạo participants cho tours
         tours.forEachIndexed { tourIndex, tour ->
             // Số người đã book (60-90% capacity)
-            val bookedSlots = ((tour.maxGroupSize * 0.6)..(tour.maxGroupSize * 0.9)).random().toInt()
+            val bookedSlots = ((tour.maxGroupSize * 0.6).toInt()..(tour.maxGroupSize * 0.9).toInt()).random()
 
             // Tạo participants
             var totalBooked = 0
